Add authentication support for Companion cameras
Companion cameras require the `Axis-Orig-Sw=true` query string parameter in the rtsp url. Otherwise they respond with a 403.

I do not have the possibility with another camera model, so it's unclear if this affects other cameras. In that case AxisDevice might need some parameter to control this or the info could be derived from Vapix.